Update index.md
Hello,

I'm an Android developer at Redfin and we've been working to incorporate Fresco in our app.  With the updated release of our app today, we will be replacing our old image loading library, and will be using Fresco to load all images in our app! Really excited for the results.

Anshu