Created AnimatedBaseSupport module
Summary: This is the first step in the process of nineoldandroid removal. This module differs from the animated-base just because it uses nineoldandroid classes. In the next diff I'm going to remove the dependency of the animated-base from the nineoldandroid classes. In the third diff I'll implement a way to use the support classes if present in the classpath. If not the classes from the animated-base will be used.

Reviewed By: kirwan

Differential Revision: D3052791

fb-gh-sync-id: cb7507bacec5e05edfad29e4dace5e4d0107fd0b
fbshipit-source-id: cb7507bacec5e05edfad29e4dace5e4d0107fd0b