Add ImagePipelineExperiments
Summary:
This creates an area within the ImagePipelineConfig where we can clearly place new config options while we experiment with them.

The intention is for these configuration options to be volatile, liable to change and sensible to avoid for open source users unless they have a strong desire otherwise.

This will allow us to move faster on such experimentation without the same worry of whether we're breaking public APIs.

Reviewed By: massimocarli

Differential Revision: D3253101

fb-gh-sync-id: c696fc4ce3e157123d8f24aa738af9ff0962c7d4
fbshipit-source-id: c696fc4ce3e157123d8f24aa738af9ff0962c7d4