Make demo build against internal modules
Summary:
Since I changed the Fresco demo to build against the remote modules it's become clear that is not intuitive. It's also created a problem for anyone who wants to demonstrate a now feature in the demo.

This change makes the internal build flavor active again while leaving the remote build flavors commented out so they are still available for testing new releases but don't get in the way of development between releases.

This will fix our current CI build failure.

Reviewed By: oprisnik

Differential Revision: D3340477

fbshipit-source-id: 06707093aa2783f5efb89c41341a2f24d438fe52