Properly handle requestDisallowInterceptTouchEvent
Summary:
If the current gesture performed on our ZoomableDraweeView haven't caused the transform to be corrected, we request our parent to not intercept events as we are going to hndle them. If on the other hand the transform got corrected, it means we tried to scroll past the bounds and we will let our parent intercept the events in case it wants to.

As a consequence we can zoom-in, pan the image around, and once we reach the edge, the parent will take over and perform paging, as the user would expect. With this fix I hope we solved issues with ZoomableDraweeView within another scrollable containers such as ViewPager, etc.
Closes https://github.com/facebook/fresco/pull/1243

Reviewed By: balazsbalazs

Differential Revision: D3360024

Pulled By: plamenko

fbshipit-source-id: 40f6fb1f2c6002a79e37e18d7e3518d29665eef6