handle WEBP_MUX_NO_BLEND properly in animated WebP
Summary:
There are some problem to process animated WebP images with NO_BLEND, sub frame.

This is the video of the problem.
https://dl.dropboxusercontent.com/u/11995554/2016_07_12_21_30_54.mp4

And this is the webp image of the problem.
https://dl.dropboxusercontent.com/u/11995554/11_21.webp

`WEBP_MUX_NO_BLEND` means that you do not need previous pixels of current frame region. But fresco ignores whole previous frame.

This is animation decoding logic in WebP repo.
https://github.com/webmproject/libwebp/blob/master/src/demux/anim_decode.c#L305
Closes https://github.com/facebook/fresco/pull/1356

Reviewed By: kirwan

Differential Revision: D3555312

Pulled By: aagnes

fbshipit-source-id: d81c277baf949a90eb8b2a23f8e76f7ec049bcee