Add ControllerViewportVisibilityListener for hints about the drawee entering/exiting the viewport
Summary: For various logging it is useful to know whether an image has been loaded while it was on screen. As Android has no definite way to get information about views entering and exiting the screen, it is the best when the client provides this information as hints. This change adds an optional `ControllerViewportVisibilityListener` to receive events from a new callback method `DraweeController#onViewportVisibilityHint`

Reviewed By: oprisnik

Differential Revision: D3655857

fbshipit-source-id: 6b9d5cd92deeaa03cccddb5a9f5704800571ce12