Added warning message if Fresco is initialized more than once
Summary:
Fresco should only be initialized once, so we should have a warning if somebody re-initializes Fresco since this causes everything to be recreated and can result in memory leaks.
We use finalizers to avoid most of these leaks but it's still something that should be avoided - especially if you're doing it unintentionally.

In the future, we plan to crash here for debug builds.

Reviewed By: kirwan

Differential Revision: D3669241

fbshipit-source-id: dd9600bdb8d1543cf8f193e76bb7f5b943206231