Update image format of encoded image if not already known
Summary: We already call `parseMetadata()` on the `EncodedImage` in various producers (which tries to set the image format) so in theory the image format should never be null in the first place. However, if we have an error and we manage to determine the image type later in the decoding step, we might as well update the encoded images so that subsequent decoding steps are faster.

Reviewed By: massimocarli

Differential Revision: D3771285

fbshipit-source-id: abcd683bd0ae0585927c944dd89ea9d6432d992a