Allow for fetching local encoded images
Summary: Before this diff, one could only fetch encoded images from the network (i.e. http/https). However, there are use-cases where fetching through the pipeline is convenient. Examples are: it's already in the encoded cache and saves us a disk access or we want to leverage the Fresco webp to jpeg transcoding. Furthermore, this makes the imagepipeline a bit more transparent to use.

Reviewed By: kirwan

Differential Revision: D3827789

fbshipit-source-id: ab34347ba4fde5dce39b7de5483be82cde08a510