Checked in code style settings for Fresco project
Summary:
This checks in the file that controls the code style when using IntelliJ or Android Studio.

This is desirable due to the Facebook style being a little non-standard (at least in terms of indentation) which results in unecessary effort to fix the formatting of PRs, either by the original author or by the importer.

Reviewed By: lambdapioneer

Differential Revision: D3822098

fbshipit-source-id: fa086c1129f3bd1de248e1f9e7f73c058cfbe457