Replace ImageRequest's auto-rotate flag with RotateOptions
Summary:
This replaces the auto-rotate flag in ImageRequest with an object allowing for more control.

Options include forcing rotation of 0, 90, 180 or 270 degrees or taking the EXIF metadata.

RotateOptions also incude an option to defer rotation until the bitmap is rendered, as drawee supports this.

ImageRequestBuilder.setAutoRotateEnabled(boolean) is now deprecated and creates default RotateOptions to use the EXIF data.

Reviewed By: lambdapioneer

Differential Revision: D3598921

fbshipit-source-id: 832f47442af749f156f95c7e713b9916628fdf8a