Create DiskCacheReadProducer to read only
Summary:
This creates a new producer by taking only the parts of DiskCacheProducer which reads from the disk cache.

This will provide the benefit of being able to place a fallback step in between write and read which may be able to read an alternative file from cache without having to force it into the existing producer.

Reviewed By: massimocarli

Differential Revision: D3886753

fbshipit-source-id: dd7cd866176398ebd64af30f5d8e61f36ccad5d8