Create DiskCacheWriteProducer to write only
Summary:
This creates a new producer by taking only the parts of DiskCacheProducer which writes to the disk cache.

This will provide the benefit of being able to place a fallback step in between write and read which may be able to read an alternative file from cache without having to force it into the existing producer.

Reviewed By: massimocarli

Differential Revision: D3886765

fbshipit-source-id: a02b58660fcaec5b78c19e22ac7067029ad7b32c