Create DiskCachePolicy
Summary:
DiskCachePolicy is a new interface intended to extract the interaction with the default and small caches from the producers and into new collaborators.

This will enable the reuse of DiskCacheReadProducer for media variations fallbacks.

For now I've created the simplest implementation to read and write from whichever cache the image request specifies.

Reviewed By: oprisnik

Differential Revision: D3937037

fbshipit-source-id: 2a78900cec680193acfa57be74686f1f5389e731