Store cache key in EncodedImage
Summary:
Adds a CacheKey member inside the EncodedImage class.

With media variations it is possible for an image to be sent back up the producer chain which should not be cached in the encoded memory cache against the request's original cache key. By storing the key in the returned image we can avoid doing this.

Reviewed By: massimocarli

Differential Revision: D3987800

fbshipit-source-id: 26345352dcf134753941536fbd58135bb71015b8