Replace images used in the media variations sample
Summary:
This diff changes the image used for the sample app to one which has a portrait orientation and with larger images for the smaller sizes.

This turned out to be necessary to test with loading smaller sizes as with the old images it was very difficult to be in a postition where the smaller sizes were big enough for the main request.

Reviewed By: lambdapioneer

Differential Revision: D4001771

fbshipit-source-id: 563a904b9293bf8bbf3f084be95c5ce26cd10a00