Allow for Bitmap.Config being null in BasePostProcessor
Summary:
Before this diff, the post processor would silently fail when it encounters a Bitmap that has no public configuration (i.e. `Config==null`). This is the case when requesting a preview image for an animated GIF.

After this diff, the post processor will fallback on using a generic `ARGB_8888` for the destination bitmap instead of creating an identical clone of the source one.

Reviewed By: oprisnik

Differential Revision: D4029653

fbshipit-source-id: e0eb4782bd4187ee9072421b0c0cd46c6ba05b6d