Mitigate crash in ForwardControllerListener caused by race condition
Summary:
In #1504 a race condition is described that leads to a `OutOfBounds` in `ForwardingControllerListener`. It happens when a `ControllerListener` is removed while an event is propagated in the `for` loop of any of those methods.

This diffs mitigates the crash by moving the `get(i)` part inside `try-catch` block. By this we still have visibility when it happens, but it does not crash the app.

Reviewed By: oprisnik

Differential Revision: D4037029

fbshipit-source-id: 6d110c2d2d2f32820579d4c998d4fcc33a41745b