Allow request to force download of primary Uri even if alternative found
Summary:
A flag has been added to MediaVariations which requests that the original image should be requested even if an alternative image was found which otherwise appears good enough to not require the network request.

This is needed (and used here) for the sample app which otherwise would not download the individual variants once one has been saved.

Reviewed By: oprisnik

Differential Revision: D4082504

fbshipit-source-id: b021567cf815dd936fa322fe4fd76479d4fa36cb