Small edit to docs - mainly for testing
Summary:
While the change here is good, this is mainly to test GitHub Pages.
We want to move the docs to master and serve from there, but it is
possible the docs ran into the bot commit issue that hampered nuclide
in the past. This is testing this.
Closes https://github.com/facebook/fresco/pull/1611

Differential Revision: D4381947

Pulled By: lambdapioneer

fbshipit-source-id: 1439a45cc2c122a25c7c7a8ea321094213f4126c