Synchronize on same object when reading or writing media variations database
Summary:
Ensure that write and reads to the index database synchronize on the same object. Previously, reads were in a synchronized method but there was no protection around writes.

Also, this removes the unnecessary `close()` call which existed after reading from the database (no close call was present in the writing case).

Fixes #1667

Reviewed By: lambdapioneer

Differential Revision: D4565894

fbshipit-source-id: 90b54c86f6f1f0b2180704bb076118b9adff47f0