Move media variations sample to showcase app
Summary:
Moves the existing media variations sample app into the Fresco showcase.

This allows the sample to benefit from features of the showcase app (such as editing settings) and removes the need for a separate app to be installed in order to try out the feature.

I'll add this sample to the menu in a following diff.

Reviewed By: lambdapioneer

Differential Revision: D4591661

fbshipit-source-id: 3b2ef5d84236440fb9d0563ce2b1f67ece0fe7ce