Avoid making multiple copies of the same list
Summary:
The existing implementation of ImmutableList.of(...) is inefficient. It makes a temporary copy of the list, then copies the contents over again into the final list. This unnecessarily uses both allocations (expensive on Android) and CPU time.
This diff makes a single copy of the data, as would be expected by an immutable list implementation aiming to manage its own data.

Reviewed By: IanChilds

Differential Revision: D4598314

fbshipit-source-id: ad194c3ba5ef35379ad8b13283474edd62fd0cd9