Don't rely on cache's memory index for media variations
Summary:
The original implementation of the MediaVariationsFallbackProducer only ran an in-memory index check for each variant.

In cases where the index is incomplete (as is common with DiskStorageCache) this results in false misses.

With this change, a full disk read is attempted for each variant until one is found. The list is sorted so that the most appropriate variant is tried first.

The media variations sample had masked this shortcoming due to its row of variants being loaded before the main image.

Reviewed By: massimocarli

Differential Revision: D4656961

fbshipit-source-id: 91314d3eabfe3b62ad786159d2cac6c3195b7262