Change media variations index to return instance of MediaVariations
Summary:
In order to hold extra data about results from queries to the index, this diff changes the return type to be an instance of `MediaVariations` instead of `List<Variant>`.

This enables following diffs which will include pulling a media ID which isn't explicitly included in the image request, as well as another which will add to the map included in producer listener callbacks from this producer.

`MediaVariations.getVariants()` is also removed and replaced by two methods: `getVariantsCount()` and `getVariant(int)` to remove any chance of the list being modified.

Reviewed By: massimocarli

Differential Revision: D4666084

fbshipit-source-id: 1cf21bcedd0c9c7162f59973dc8c211ac21103db