Add source property to MediaVariations and include in listener events
Summary:
Adds a new property to the MediaVariations class to distinguish between occasions when MediaVariationsFallbackProducer finds an alternative that was explictly listed or one which was found in the index thanks to a shared media ID.

This is also added to listener events from that producer.

A following change will introduce another potential source.

Reviewed By: massimocarli

Differential Revision: D4666715

fbshipit-source-id: 0e5b61972fa57f4ae7703b417f67298a6c67cc6f