Revert: Build jpegturbo as a shared library
Summary: This reverts pull-request #1686 (https://github.com/facebook/fresco/pull/1686). After it has been landed it was discovered that it would increase the APK size for apps that just use the *core* library. We are committed to keep the core small and therefore we decided to revert it to the previous state.

Reviewed By: erikandre

Differential Revision: D4827270

fbshipit-source-id: b59ce42058078b766d3b75bb8f2f27d1e7b06dc8