Add ProducerListener.onUltimateProducerReached()
Summary: Adds a new callback to ProducerListener which is called by a producer if, and only if, it's responsible for the ultimate satisfaction of the request. That's to say, if the image is found in the encoded memory cache, `EncodedMemoryCacheProducer` will call this listener method. Likewise for other cache producers, `NetworkFetchProducer` and extensions of `LocalFetchProducer`.

Reviewed By: oprisnik

Differential Revision: D4779814

fbshipit-source-id: f24e7165d84c888fc6739320979125b20ce9f274