Added onFramePrepared to BitmapFrameCache
Summary:
This change adds a `cachePreparedFrame` to `BitmapFrameCache`, which is called for frames that are available ahead of time and have not yet been rendered.
This is necessary since we don't want to immediately recycle these frames until they have at least been drawn once.

Actual implementations for this will be added separately. Right now, none of them is able to cache these frames.

Reviewed By: lambdapioneer

Differential Revision: D4875544

fbshipit-source-id: ce3f676d716f494a2c8b92dde0327f9c6538987d