Better FrescoFrameCache bitmap reference conversion
Summary:
The current approach for converting a `CloseableReference<CloseableImage>` to a `CloseableReference<Bitmap>` has a flaw:
We constantly do conversions in both directions to read / write from the cache. This means we create new ResourceReleasers for one direction and  wrap the same `CloseableReference<Bitmap>` in new `CloseableStaticBitmap` objects.
This means, that we always increase the call stack and keep references to all previous `CloseableStaticBitmap`. This will eventually result in stack overflow exceptions and it wastes memory.

The alternative solution here is to use a clone of the underlying Bitmap reference instead in order to be able to immediately close old `CloseableStaticBitmap`s and to get rid of custom `ResourceReleaser`s.

Reviewed By: kirwan

Differential Revision: D4921775

fbshipit-source-id: 93adc493dc24989c8f47a31727ba7fc8a882d1b0