Don't trigger a request failure if resizing fails on an incomplete response
Summary:
It was possible for the first response to hit the first incomplete response to hit the ResizeAndRotateProducer, fail to load the image data and fail the entire request, even though when the next piece of data arrived it ran successfully.

A later successful completion of the request was ignored.

With this diff, that initial failure is ignored so that the next one may succeed.

Reviewed By: erikandre

Differential Revision: D5052202

fbshipit-source-id: 33f779fce172baa90ce0d6122f664b3df5f161e1