Create new version of onNewResult with Status flags
Summary:
In order to allow more information to be passed, a new version of `onNewResult()` is added which receives an int called `status`.

For now, this lives alongside the existing `onNewResult(T, boolean)` until all consumers have been edited to implement the new one. This also means `BaseConsumer` temporarily contains implementations of both methods. When complete the old one will be removed and the new one will become abstract.

Reviewed By: erikandre

Differential Revision: D5060836

fbshipit-source-id: 1c16b4738ecb55f9b766e72f6e855c9a11b810ca