Catch exception when opening an ExifInterface for a non-jpeg
Summary:
Fixes #1558 and #1760.

When a local file is loaded an attempt is made to load the thumbnail from its exif data but that is only available for JPEGs.

Trying the same with a PNG lead to an `IOException` which wasn't being caught within the producer. Catching it allows us to return null which is the appropriate thing to do if no thumbnail is available.

Reviewed By: erikandre

Differential Revision: D5077525

fbshipit-source-id: fdfd8e0bc369e4672729c6e3bfde5b035f836149