Stop reading encoded cache key from EncodedImage
Summary:
Caching using keys held by `EncodedImage` was a work-around to avoid caching alternative from media variations using the originally-requested URL.

Now that we can stop the image being cached at all in that case, we can go back to using the original cache key.

I've also deprecated the setting and retrieving of the cache key in `EncodedImage` but haven't removed it as it may be used by consumers of Fresco.

Reviewed By: erikandre

Differential Revision: D5077826

fbshipit-source-id: 25aafa6d9127016b04b1fb48fe815c1d2000c12c