Bitmap.prepareToDraw: Add BitmapPrepareProducer to ProducerSequenceFactory
Summary: This adds the `BitmapPrepareProducer` to the `ProducerSequenceFactory`. It will be added at the very end of the pipeline (after the optional post-processor) since we are interested to upload the final Bitmap that is passed back to Drawee. Also note that we do not add the producer for versions below Android N and avoid the overhead this way.

Reviewed By: erikandre

Differential Revision: D5106621

fbshipit-source-id: 6c76551f1c8dd6e97837b3846197a1ffda501dd8