Add Showcase sample for partial loading
Summary:
Adds a sample to the Showcase app in which we can load part of an image and then follow that with a request for the remainder of the image.

The first "load partial" button loads just the first 30000 bytes of the image and the second launches a request to load the remainder of the image.

Currently, the first image is cached against the URI so the rest of the image isn't loaded. This enables the testing of the implementation as it's fleshed out in following diffs.

Reviewed By: erikandre

Differential Revision: D5120995

fbshipit-source-id: 38630c8fa273ba341c32b3e0c66bf37514d4f815