Moved DrawableFactory to imagepipeline-base
Summary:
NOTE: This is a breaking change for custom drawable factories (e.g. if you add Keyframes / SVG support).
In order to fix the build, change the `DrawableFactory` import from `import com.facebook.drawee.backends.pipeline.DrawableFactory;` to `import com.facebook.imagepipeline.drawable.DrawableFactory;`.

`DrawableFactory` is closely related to the imagepipeline since it takes a `CloseableImage` (which lives in `imagepipeline-base`) and creates the Drawable that renders the image.
By moving it to `imagepipeline-base`, we have everything required to add support for new image formats in the same module and it's also possible to use `DrawableFactories` standalone without the need to depend on Drawee.

Reviewed By: lambdapioneer

Differential Revision: D5255363

fbshipit-source-id: cf6ea0c217c1244acdf09abdbd73487aa78bf9ed