Fix IncreasingQualityDataSource to keep low res image if hi res fails
Summary:
Turns out that if the high res image fails to load (e.g. instant 404), while the low res is still loading, the low res is cancelled. This diff makes it so that IncreasingQualityDataSource stores the failure cause for the high res while waiting for all other requests to complete.

Related:

* https://github.com/facebook/fresco/issues/1676
* https://github.com/facebook/fresco/pull/1677

Reviewed By: oprisnik

Differential Revision: D5397825

fbshipit-source-id: 95fb9651f016fb38463f3a085366af03a5d33a32