Merging PR #1825 - Updating all documentation
Summary:
Updates the http://frescolib.org/docs/ with a new version that fixes existing issues and adds new content. Some content has been re-prioritized to make it easier to consume for new users of the Fresco library.

Many of the documents now have images and videos that demonstrate the changes. Consistent links to the new showcase app make it easier to find code to start with.

This PR contains the work of multiple authors:
 - erikandre
 - foghina
 - kirwan
 - lambdapioneer

Reviewed By: foghina

Differential Revision: D5442768

Pulled By: lambdapioneer

fbshipit-source-id: f0b2885c6cc8790abf8fda539957027d5f805fec