fix BetterImageSpan size calculation
Summary: BetterImageSpan getSize is supposed to increase the top or bottom of the font metrics to fit the image but it doesn't increase the bottom. If the image is taller than the font size then it will be cut off from the bottom. This fixes the size calculation to also increase the font bottom metric.

Reviewed By: oprisnik

Differential Revision: D6131715

fbshipit-source-id: eaad38f77700d19842a3c5b161999afd4b9961c2