Support inverted rotation orientations
Summary:
When an inverted EXIF orientation is detected, it executes a different new JpegTranscoder method that accepts the actual raw EXIF value rather than rotation angle in degrees. Rotation angle is set to 0 for inverted orientations.

This only rotates inverted orientations in the pipeline and does not yet support it for deferred rotations. Normal orientations can still be deferred.

Reviewed By: kirwan

Differential Revision: D6267963

fbshipit-source-id: 8a358dbfd5357a93cacb6265fe77b09d08b17975