ZoomableDraweeView: fix missing transformation callback if image loading is slow
Summary:
`DefaultZoomableController` exposes a listener for transformation changes, `Listener#onTransformChanged(Matrix transform)`.
However, the listener is currently only invoked if the controller is enabled. This works fine if the image has been loaded already and anything else triggers another transformation change.
If image loading is slow (e.g. if it's a big image that is not in memory already) and there is no other transformation change (e.g. bounds change, listeners will miss the final transformation event since the controller is still disabled when the final image is set. By simply enabling the controller before the computations for the final image are performed, `DefaultZoomableController` will properly notify listeners.

Reviewed By: foghina

Differential Revision: D6469214

fbshipit-source-id: 47be29bd6740804d5a04b423f2561f6ae4c05e74