Do not treat response as partial with Content-Range if range includes…
Summary:
Do not treat the response as partial when "Content-Range" header is set if range includes the entire message.

If the header "Content-Range" is ever passed in from a network request, the image will not be saved to cache. What I have changed is that, if "Content-Range" includes the entire content range for the image, then do not treat as a partial image.

I have tested this on my personal repo by trying to fetch an image with "Content-Range" set, but I have not created any unit tests.
Closes https://github.com/facebook/fresco/pull/2018

Reviewed By: foghina

Differential Revision: D6682650

Pulled By: erikandre

fbshipit-source-id: cd25c6e18066240a10db7b8fcb5b8f08457dc2fc