Added basic region decoding sample
Summary:
This adds a very simple region decoding sample to the Showcase ap.
This is not fully working yet. It only works for JPEGs and if the selected image is resized or downsampled, the decoded region will be wrong and you will see a black area around the actual region.

Reviewed By: erikandre, foghina

Differential Revision: D6759270

fbshipit-source-id: 576310f74ed85ed27a1303642ddd1cd1285d23de