Default request timeout
Summary:
The motivation for this change comes from a problem in an app I'm working on, under very bad connectivity there was an issue where timeout meant the image wasn't being downloaded. After lots of testing and consideration to swap to a regular ImageView and download a Bitmap, I investigated further and found under the worst conditions it was taking 20 seconds to download the image.

What existing problem does the pull request solve?
It solves a timeout issue when downloading images over a weak or poor connection.

Unit test added to confirm connection timeout is set.
Closes https://github.com/facebook/fresco/pull/2068

Reviewed By: erikandre

Differential Revision: D7288817

Pulled By: oprisnik

fbshipit-source-id: 6a306ac09760dcd6ed803a72262d905bd5dd62f8