Added ImageOrigin and ImageOriginListener for Drawee controllers
Summary:
This change adds a new ImageOrigin enum and ImageOriginListener to listen for the origin of a given image, which can be any of our caches, network, or unknown.
In this change, the origin listener is only triggered when the image is served directly from DraweeController's memory cache without going to the pipeline. This will be expanded in the future to work for all origins.

Differential Revision: D7394707

fbshipit-source-id: 696d3789bec10d86a00f4da999e4ef95e4f69a68