Remove unobserved max_memory_to_use flag
Summary:
The `max_memory_to_use` flag is not used by the libjpeg-turbo version shipping with Fresco (this is 1.5.0). It starts to be obeyed from libjpeg version 1.5.2 and above.

Removing the flag to allow the update to 1.5.2 match the existing behaviour. The current value would prevent displaying any progressive JPEG >3000x2500px.

Reviewed By: erikandre

Differential Revision: D7533437

fbshipit-source-id: f6dc5406285277196c0134ab6c4debb59ac22b02