Allow external request listeners to be passed when fetching an image
Summary:
Currently, all image request listeners have to be set for the `ImageRequest` using an `ImageRequestBuilder`. Since `ImageRequest is immutable, it's not easily possible to add additional (internal) listeners to for image requests without cloning the request.

Now, it's possible to pass an additional listener to the `ImagePipeline` when an image is fetched.

Reviewed By: foghina

Differential Revision: D7383204

fbshipit-source-id: 7d59db31059f4f2bd915cd89f022ba72d20fc1c0