Move showcase app recyclerview example from loremipsum.com to picsum.photos
Summary:
The previous service is no longer functioning properly. Changing the recyclerview example to use https://picsum.photos/ instead. It's also easier as we don't have to care about categories.

To have reproducible results, we use image ids on the URI. We have them in a "fixed random order" by using a `Random` with fixed seed. We need to randomize as the dataset is sorted (e.g. the first 10 images all feature laptops).

Some images will error with 404 (e.g. that particular ID is no longer part of the dataset). But I think this is actually nice as it tests the error handling paths.

Reviewed By: erikandre

Differential Revision: D7774124

fbshipit-source-id: 5511dae767c9b4db018f7e802d1aa3f3fefb90e5