Pass DraweeController down to data source generation
Summary:
NOTE: This is a breaking Fresco change if you have custom Drawee controller builders. Please update your method signatures to include the `DraweeController` as the first parameter.

Currently, data source generation is independent of the controller, which limits our possibilities. This commit passes the DraweeController down to the data source so that we can access it (e.g. to attach request listeners to the request).

Reviewed By: foghina

Differential Revision: D7860926

fbshipit-source-id: 492724824a02fa84324473b1ac87452b0c34d2d5