Hook up image perf logging framework with DraweeController
Summary:
NOTE: This is a breaking change. If you implement custom DraweeControllers, please update your implementations to match the updated method signatures.

This hooks up the perf logging framework.
We need to pass the controller down in order for the correct listeners to be attached when an image is requested by the data source.

Reviewed By: foghina

Differential Revision: D7673132

fbshipit-source-id: 3717de484dd23ec020f9ea024c1a83ab4ff674d9