Support Re-using Bitmaps with RGBA_F16
Summary:
If the Bitmap Config is RGBA_F16 it uses 8 bytes per pixel when computing the bitmap size. And if the image is wide gamut but the bitmap config does not use RGBA_F16, the size of the bitmap is manually computed to ensure that the requested size is accurate.

I created a new OreoDecoder to differentiate bitmap size computation and removed code duplication in decoders.

Reviewed By: oprisnik

Differential Revision: D9119486

fbshipit-source-id: 8d162dfc0253789c690fe2677ac2f116b1981bef