BaseBitmapReferenceDataSubscriber: Clone reference instead of detaching
Summary: For data source subscribers, Fresco passes down a `CloseableReference<CloseableImage>`. Detaching the reference for the closeable image in case it is a `CloseableStaticBitmap` will detach the bitmap from the cached image since the same `CloseableStaticBitmap` is returned every time instead of a cloned one. This means that we must clone the underlying bitmap reference instead of detach. Otherwise, subsequent calls will fail.

Reviewed By: defHLT

Differential Revision: D9786558

fbshipit-source-id: 0c07bc0c413ba438113f10ad01bd0baa217f082e