fix bug when using DownSample and RegionDecoder together (#2203)
Summary:
To fix the bug discribed in issue [2201](https://github.com/facebook/fresco/issues/2201)

The Image can not show correctly when using Resize and RegionDecoder together with `downsample` setting to `true`. Here I change the `targetWidth` and `targetHeight` with region decoder to fix the bug.
whe `setDownsampleEnable` is not  set to `true` or using default value. Resize is a process of pipeline before decoding (only for JPEG). so the region is not the same with the former. but this can be set in custom decoder.
Pull Request resolved: https://github.com/facebook/fresco/pull/2203

Reviewed By: oprisnik

Differential Revision: D9677783

Pulled By: defHLT

fbshipit-source-id: c010728047ca4f8a876cb9284ee9efebcbc97ab1