Proper support for GIFs when using a RetainingDataSource (#2185)
Summary:
During the development of the Giphy app we choose to use fresco as our main image library. The application loads different types of image renditions, starting from low-quality images to high-quality ones. We wanted to be able to change the quality of a GIF displayed in a `DraweeView`.

When making a simple request to change the resource of a `DraweeView`, fresco will clear the current content, display the placeholder, then load the new resource. Even with the resources preloaded, there is a noticeable flickering.

Available solutions:
1. Use a low-res/hi-res schema for loading images. Unfortunately, in this schema, the low-res image is not animated.
2. Use a `RetainingDataSource`. While it's purpose was exactly what we were looking for, it turns out that this data source, does it's magic by lying to the `AbstractDataSource` code and keeping it in a forever `PROGRESS` state. That means, the datasource will never get to notify the listening controllers. This is the reason GIFs are not playing and controller callbacks are not fired using this `DataSource`.

Given a list of GIFs, by using a `RetainingDataSource` we should be able to replace the current playing view with a freshly loaded one, without the user notice any flickering or blank spaces between the 2 resouces.

I've created a set of test GIFS in the `ImageUriProvider` and I updated the `RetainingDataSourceSupplierFragment` to use those resouces, in order to demonstrate the new achieved behaviour.

**Demo video:**
https://youtu.be/A67X6Vn9VjY
Pull Request resolved: https://github.com/facebook/fresco/pull/2185

Reviewed By: erikandre

Differential Revision: D9446858

Pulled By: oprisnik

fbshipit-source-id: dd30f00818eb7ab80b468e0a1ed243a700c51ae3