Disable CircleCI instrumentation test with network dependency
Summary:
Triaged this down to a network issue, as I can reproduce the same behaviour locally by turning on flightmode.

I suspect that it is due to the change to SSL on our side. For now, I'll disable the test that's failing. But plan to follow up with a fix once I have more time

Reviewed By: erikandre

Differential Revision: D10083556

fbshipit-source-id: 2720a5fb38403a23b9398abab85e1d945d342b33