Simplifies enabling Fresco non-native ImageTranscoder in ImagePipelineConfig
Summary:
Simplifies constructing an `ImagePipelineConfig` with the java-only `ImageTranscoder` by passing an `ImageTranscoderType` instead of having to construct the object.
The current implementation was too complicated and tend to push clients to disable only the native image transcoder and keep native memory allocation by writting:
```
ImagePipelineConfig config =
    ImagePipelineConfig.newBuilder(context)
        .experiment()
        .setNativeCodeDisabled(true)
        .setMemoryChunkType(MemoryChunkType.NATIVE_MEMORY)
        .build();
```
This version was easier to write as it didn't require to construct a `SimpleImageTranscoderFactory` and having to determine the parameters to pass, but it didn't feel natural.
This diff simplifies this process and clients can now write it like this:
```
ImagePipelineConfig config =
    ImagePipelineConfig.newBuilder(context)
        .setImageTranscoderType(ImageTranscoderType.JAVA_TRANSCODER)
        .build();
```

Reviewed By: oprisnik

Differential Revision: D10255443

fbshipit-source-id: e17d034c3ee06526e6ebe9e11e9bbd4bf1518f33