Converted ShowcaseApplication and ImageUriProvider to Kotlin
Summary: This simplifies how we use `ImageUriProvider` in Showcase. It's created once within the application and can then be used in all samples instead of being a singleton and depending on the given Context. `ImageUriProvider` has been converted with Android Studio and slightly adapted. I'll make optimizations and clean it up later.

Reviewed By: defHLT

Differential Revision: D15065902

fbshipit-source-id: c709e9b3d85edce230110789331b6231d9de269e